requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
passlib[bcrypt]
pyjwt
